+++ NEWS.md
@@ -3,7 +3,12 @@
 ## Documentation
 
  * Drop duplicated arguments from `help("parse.SmartComments")`.
- 
+
+## Deprecated and Defunct
+
+ * S3 method `warnings()` for `Verbose`, and the corresponding generic
+   function, has been deprecated in favor of the `printWarnings()`.
+
 
 # Version 2.12.1 [2022-10-30]